♻️ refactor(monitoring): split data adapters and meta service
Move monitoring chart builders, analytics adapters, event row conversion, and metadata loading into dedicated model and service modules.

This leaves useMonitoringData focused on React state orchestration while preserving its existing return contract.
